+++ spec/rubocop/cop/rails/before_destroy_spec.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+RSpec.describe RuboCop::Cop::Rails::BeforeDestroy do
+  subject(:cop) { described_class.new(config) }
+
+  let(:config) { RuboCop::Config.new }
+
+  # TODO: Write test code
+  #
+  # For example
+  it 'registers an offense when using `#bad_method`' do
+    expect_offense(<<~RUBY)
+      bad_method
+      ^^^^^^^^^^ Use `#good_method` instead of `#bad_method`.
+    RUBY
+  end
+
+  it 'does not register an offense when using `#good_method`' do
+    expect_no_offenses(<<~RUBY)
+      good_method
+    RUBY
+  end
+end